+++ tag_nuke.php
@@ -4,20 +4,30 @@ include('cd.php');
 ini_set('max_execution_time', '3600');
 $CurrentUser = Authentication::Authenticate();
 
-$Tags = Tag::GetTags();
-
-/* @var $t Tag */
-foreach($Tags as $t)
+if($CurrentUser->hasPermission(RIGHT_TAG_CLEANUP))
 {
-	$t2as = Tag2All::GetTag2Alls(sprintf('tag_id = %1$d', $t->getID()));
+	$Tags = Tag::GetTags();
 	
-	if(!$t2as){
-		Tag::DeleteTag($t, $CurrentUser);
+	/* @var $t Tag */
+	foreach($Tags as $t)
+	{
+		$t2as = Tag2All::GetTag2Alls(sprintf('tag_id = %1$d', $t->getID()));
+		
+		if(!$t2as){
+			Tag::DeleteTag($t, $CurrentUser);
+		}
 	}
+	
+	$infoSuccess = new Info($lang->g('MessageTagsCleaned'));
+	Info::AddInfo($infoSuccess);
 }
-
-$infoSuccess = new Info($lang->g('MessageTagsCleaned'));
-Info::AddInfo($infoSuccess);
+else
+{
+	$e = new Error();
+	$e->setErrorNumber(RIGHTS_ERR_USERNOTALLOWED);
+	$e->setErrorMessage(Error::TranslateError(RIGHTS_ERR_USERNOTALLOWED));
+	Error::AddError($e);
+}
 
 HTMLstuff::RefererRedirect();
